Added an outDir to the tsconfig.json
I added an outDir so all of the output of the TypeScript and JavaScript files get emitted into `.svelte-kit`. I also added a exclude parameter so the output of these files do not get redundantly transpiled, and allows tsconfig to be read when an exception occurs.

This solves a heuristic parsing error for reading the tsconfig.json file ('Parsing error: Cannot read file '.../tsconfig.json'')  that I believe is specific to VS Code users. I did not get this error in Vim. The tsconfig.json located in `.svelte-kit` should take priority for configuration for the project, however VS code sometimes likes to think the tsconfig in the parent directory should take priority. 

The error can show up before building, or when adding new components or files if an outDir and exclude is not made. Saved a bit of headache for me! :)